Adding pcpush and pcpop commands + refactoring
This commit adds implementations of pcpush and pcpop commands. The
names are selected in this way because the capabilities created and used
by these commands live on the same stack as the pini-push/pop commands.
The well-scopedness of using these is important and is enforced by the
runtime; otherwise the reasonable invariants of the call stack may be
violated.

Observe that pcpop traverses the stack at runtime and declassifies the
pc-levels of all stack frames for as long as needed (after checking that
the provided authority is sufficient). Observe also that the check
takes place just once, because we rely on the monotinicity of the levels
on the frame here.

As part of the implementation, we have also refactored the Thread
module.