Adding pcpush and pcpop commands + refactoring
This commit adds implementations of pcpush and pcpop commands. The
capabilities created/used by these commands live on the same stack as
the capbilities created by pini-push/pop commands. The runtime enforces
that all these capabilities are used in a well-scoped manner.  Otherwise
the reasonable invariants of the call stack can be violated.

Observe that pcpop traverses the stack and declassifies the pc-levels of
all stack frames for as long as needed (after checking that the provided
authority is sufficient). Observe also that the authority check takes
place just once, because we rely on the monotinicity of the levels on
the call stack.

As part of this commit, we have also refactored the Thread module.