Force color output in warnings and errors
Google Actions do not execute in a TTY which leads to colors not being emitted properly for tests that have errors and warnings. We now use a separate console object that forces color output. This can potentially introduce options for testing with nonTTY terminals when that becomes relevant. We will need to in the future parametrize this option and potentially use the right parameters in the golden framework in GA.